requirements.txt
asgiref==3.4.1
astroid==2.8.0
attrs==21.2.0
autopep8==1.5.7
coverage==5.5
Django==3.2.7
djangorestframework==3.12.4
Faker==8.14.1
iniconfig==1.1.1
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mixer==7.1.2
packaging==21.0
platformdirs==2.3.0
pluggy==1.0.0
py==1.10.0
pycodestyle==2.7.0
PyJWT==2.1.0
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==2.12.1
pytest-django==4.4.0
pytest-sugar==0.9.4
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.2
wrapt==1.12.1
